Add completedTask function and test
